form component for ingredients and instructions
