+++ screens/AddRecipe.tsx
@@ -1,178 +1,19 @@
-import React, { useState } from 'react';
-import {
-  View,
-  Text,
-  TextInput,
-  ScrollView,
-  TouchableOpacity,
-  StyleSheet,
-  Image,
-} from 'react-native';
-import { launchImageLibrary } from 'react-native-image-picker';
-import { Recipe, Ingredient, IngredientGroup, InstructionGroup } from '../models/Recipe';
-import RecipeStore from '../store/RecipeStore';
+import React from 'react';
+import { View } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
-import Ionicons from 'react-native-vector-icons/Ionicons';
-import { useTheme } from '../hooks/useTheme';
-import Colors from '../constants/Colors';
 import Header from '../components/Header';
-import CustomPopup from '../components/CustomPopup';
-import TextInputPopup from '../components/TextInputPopup';
+import RecipeForm from '../components/RecipeForm';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { NestableScrollContainer } from 'react-native-draggable-flatlist';
 import ContentWrapper from '../components/ContentWrapper';
-import { CURRENT_SCHEMA_VERSION } from '../models/RecipeMigrations';
-import GroupsEditor, { GroupDraft } from '../components/GroupsEditor';
-
-function generateId(): string {
-  return `${Date.now()}_${Math.random().toString(36).substr(2, 9)}`;
-}
+import RecipeStore from '../store/RecipeStore';
+import { Recipe } from '../models/Recipe';
 
 export default function AddRecipe() {
   const router = useNavigation();
-  const [name, setName] = useState('');
-  const [imageUri, setImageUri] = useState<string | null>(null);
-  const [ingredientGroups, setIngredientGroups] = useState<GroupDraft[]>([
-    { id: generateId(), title: '', items: [] }
-  ]);
-  const [instructionGroups, setInstructionGroups] = useState<GroupDraft[]>([
-    { id: generateId(), title: '', items: [] }
-  ]);
-  const [sourceUrl, setSourceUrl] = useState('');
-  const [cookingTime, setCookingTime] = useState('');
-  const [calories, setCalories] = useState('');
-  const [newTag, setNewTag] = useState('');
-  const [tags, setTags] = useState<string[]>([]);
-  const [showPopup, setShowPopup] = useState(false);
-  const [popupConfig, setPopupConfig] = useState({
-    title: '',
-    message: '',
-    buttons: [] as Array<{ text: string; onPress: () => void }>,
-  });
-  const [isScrollEnabled, setIsScrollEnabled] = useState(true);
 
-  type SectionKey = 'ingredients' | 'instructions';
-  const [itemEditor, setItemEditor] = useState<{
-    visible: boolean;
-    mode: 'add' | 'edit';
-    section: SectionKey;
-    groupId: string | null;
-    itemId?: string;
-    initialText: string;
-    isHeader?: boolean;
-  }>({ visible: false, mode: 'add', section: 'ingredients', groupId: null, initialText: '' });
-
-  const findGroupsBySection = (section: SectionKey) => section === 'ingredients' ? ingredientGroups : instructionGroups;
-  const setGroupsBySection = (section: SectionKey, next: GroupDraft[]) => {
-    if (section === 'ingredients') setIngredientGroups(next); else setInstructionGroups(next);
-  };
-
-  const handleOpenAddItem = (section: SectionKey, groupId?: string) => {
-    // In singleGroup mode, we no longer expose add UI; keep handler for future or programmatic use
-    const groups = findGroupsBySection(section);
-    const targetGroupId = groupId || groups[0]?.id || null;
-    if (!targetGroupId) return;
-    setItemEditor({ visible: true, mode: 'add', section, groupId: targetGroupId, initialText: '', isHeader: false });
-  };
-
-  const handleOpenEditItem = (section: SectionKey, groupId: string, itemId: string, currentText: string, isHeader?: boolean) => {
-    setItemEditor({ visible: true, mode: 'edit', section, groupId, itemId, initialText: currentText, isHeader: !!isHeader });
-  };
-
-  const handleConfirmItemEditor = (text: string, type: 'item' | 'header') => {
-    const { section, groupId, mode, itemId } = itemEditor;
-    const groups = findGroupsBySection(section);
-    const idx = groups.findIndex(g => g.id === groupId);
-    if (idx === -1) { setItemEditor(prev => ({ ...prev, visible: false })); return; }
-    const group = groups[idx];
-
-    const items = [...group.items];
-    if (mode === 'add') {
-      items.push({ id: generateId(), text: text.trim(), isHeader: type === 'header' });
-    } else if (mode === 'edit' && itemId) {
-      const i = items.findIndex(it => it.id === itemId);
-      if (i !== -1) items[i] = { ...items[i], text: text, isHeader: type === 'header' } as any;
-    }
-    const next = [...groups];
-    next[idx] = { ...group, items };
-    setGroupsBySection(section, next);
-    setItemEditor(prev => ({ ...prev, visible: false }));
-  };
-
-  // const colorScheme = useColorScheme() ?? 'light';
-  // const colors = Colors[colorScheme];
-  const { colors, colorScheme } = useTheme();
-
-  const handleAddImage = async () => {
-    const result = await launchImageLibrary({
-      mediaType: 'photo',
-      quality: 0.8,
-    });
-
-    if (!result.didCancel && result.assets && result.assets[0].uri) {
-      setImageUri(result.assets[0].uri);
-    }
-  };
-
-  const handleAddTag = () => {
-    if (newTag.trim()) {
-      setTags([...tags, newTag.trim()]);
-      setNewTag('');
-    }
-  };
-
-  const handleDeleteTag = (tagToDelete: string) => {
-    setTags(tags.filter(tag => tag !== tagToDelete));
-  };
-
-  const toIngredientGroups = (drafts: GroupDraft[]): IngredientGroup[] => {
-    return drafts.map((g) => new IngredientGroup({
-      title: (g.title || '').trim() || undefined,
-      items: g.items.filter(it => !it.isHeader).map(it => new Ingredient((it.text || '').trim())).filter(ing => ing.name)
-    }));
-  };
-
-  const toInstructionGroups = (drafts: GroupDraft[]): InstructionGroup[] => {
-    return drafts.map((g) => new InstructionGroup({
-      title: (g.title || '').trim() || undefined,
-      items: g.items.filter(it => !it.isHeader).map(it => (it.text || '').trim()).filter(Boolean)
-    }));
-  };
-
-  const handleSave = () => {
-    if (!name.trim()) {
-      setPopupConfig({
-        title: 'Missing Information',
-        message: 'Please enter a recipe name',
-        buttons: [
-          {
-            text: 'OK',
-            onPress: () => setShowPopup(false),
-          },
-        ],
-      });
-      setShowPopup(true);
-      return;
-    }
-
-    const ingredientsGroupsOut = toIngredientGroups(ingredientGroups);
-    const instructionGroupsOut = toInstructionGroups(instructionGroups);
-
-    const recipe = new Recipe({
-      name: name.trim(),
-      imageUri,
-      ingredients: [],
-      instructions: [],
-      sourceUrl: sourceUrl.trim() || undefined,
-      cookingTime: cookingTime.trim() || undefined,
-      calories: calories.trim() || undefined,
-      tags,
-      schemaVersion: CURRENT_SCHEMA_VERSION,
-      ingredientsGroups: ingredientsGroupsOut,
-      instructionGroups: instructionGroupsOut,
-    });
-
-    RecipeStore.addRecipe(recipe);
+  const handleSave = async (recipe: Recipe) => {
+    await RecipeStore.addRecipe(recipe);
     router.goBack();
   };
 
@@ -182,348 +23,10 @@ export default function AddRecipe() {
         <Header title="Add recipe" />
         <NestableScrollContainer style={{ flex: 1 }}>
           <ContentWrapper>
-            <View style={[styles.container, { backgroundColor: colors.background }]}>
-              <View style={styles.form}>
-                <Text style={[styles.sectionTitle, { color: colors.text }]}>Name</Text>
-                <TextInput
-                  style={[styles.input, {
-                    borderColor: colors.inputBorder,
-                    color: colors.text,
-                  }]}
-                  placeholder="e.g. Spaghetti Carbonara"
-                  placeholderTextColor={colors.placeholderText}
-                  value={name}
-                  onChangeText={setName}
-                />
-
-                <Text style={[styles.sectionTitle, { color: colors.text }]}>Image</Text>
-
-                {imageUri ? (
-                  <View style={[styles.imagePreview, { backgroundColor: colors.placeholderBackground }]}>
-                    <Image source={{ uri: imageUri }} style={styles.previewImage} />
-                  </View>
-                ) : null}
-
-                <TouchableOpacity
-                  style={[styles.imageButton, { backgroundColor: colors.tint }]}
-                  onPress={handleAddImage}
-                >
-                  <Text style={[styles.imageButtonText, { color: colors.background }]}>
-                    {imageUri ? 'Change Image' : 'Select Image'}
-                  </Text>
-                </TouchableOpacity>
-
-                <GroupsEditor
-                  title="Ingredients"
-                  groups={ingredientGroups}
-                  onChange={setIngredientGroups}
-                  placeholderNewGroup="e.g. Sauce"
-                  placeholderItem="e.g. 200g tomatoes"
-                  onEditItemRequest={(groupId, itemId, currentText, isHeader) => handleOpenEditItem('ingredients', groupId, itemId, currentText, isHeader)}
-                  singleGroup
-                />
-
-                <GroupsEditor
-                  title="Instructions"
-                  groups={instructionGroups}
-                  onChange={setInstructionGroups}
-                  placeholderNewGroup="e.g. Sauce"
-                  placeholderItem="e.g. Sauté onions until soft"
-                  onEditItemRequest={(groupId, itemId, currentText, isHeader) => handleOpenEditItem('instructions', groupId, itemId, currentText, isHeader)}
-                  singleGroup
-                />
-
-                <Text style={[styles.sectionTitle, { color: colors.text }]}>Details</Text>
-                <View style={styles.detailsRow}>
-                  <View style={styles.detailInput}>
-                    <Text style={[styles.detailLabel, { color: colors.text }]}>Cooking Time</Text>
-                    <TextInput
-                      style={[styles.input, {
-                        borderColor: colors.inputBorder,
-                        color: colors.text,
-                        marginBottom: 0,
-                      }]}
-                      placeholder="e.g. 30 min"
-                      placeholderTextColor={colors.placeholderText}
-                      value={cookingTime}
-                      onChangeText={setCookingTime}
-                    />
-                  </View>
-                  <View style={styles.detailInput}>
-                    <Text style={[styles.detailLabel, { color: colors.text }]}>Calories</Text>
-                    <TextInput
-                      style={[styles.input, {
-                        borderColor: colors.inputBorder,
-                        color: colors.text,
-                        marginBottom: 0,
-                      }]}
-                      placeholder="e.g. 250 kcal"
-                      placeholderTextColor={colors.placeholderText}
-                      value={calories}
-                      onChangeText={setCalories}
-                      // keyboardType="numeric"
-                    />
-                  </View>
-                </View>
-
-                <Text style={[styles.detailLabel, { color: colors.text }]}>Tags</Text>
-                <View style={styles.ingredientInput}>
-                                    <TextInput
-                     style={[styles.input, {
-                       flex: 1,
-                       marginBottom: 0,
-                       borderColor: colors.inputBorder,
-                       color: colors.text,
-                       height: 48,  // Fixed height to match EditRecipe
-                     }]}
-                     placeholder="e.g. vegetarian"
-                     placeholderTextColor={colors.placeholderText}
-                     value={newTag}
-                     onChangeText={setNewTag}
-                     onSubmitEditing={handleAddTag}
-                   />
-                  <TouchableOpacity
-                    style={[styles.addButton, { backgroundColor: colors.tint }]}
-                    onPress={handleAddTag}
-                  >
-                    <Ionicons name="add" size={24} color={colors.background} />
-                  </TouchableOpacity>
-                </View>
-                <ScrollView
-                  horizontal
-                  showsHorizontalScrollIndicator={false}
-                  style={styles.tagsScrollContainer}
-                  contentContainerStyle={styles.tagsContainer}
-                >
-                  {tags.map((tag) => (
-                    <View key={tag} style={[styles.tagContainer, { backgroundColor: colors.tint }]}> 
-                      <Text style={[styles.tagText, { color: colors.background }]}>{tag}</Text>
-                      <TouchableOpacity
-                        style={styles.tagDeleteButton}
-                        onPress={() => handleDeleteTag(tag)}
-                      >
-                        <Ionicons
-                          name="close-circle"
-                          size={24}
-                          color={colors.background}
-                          style={styles.tagDeleteIcon}
-                        />
-                      </TouchableOpacity>
-                    </View>
-                  ))}
-                </ScrollView>
-
-                <Text style={[styles.detailLabel, { color: colors.text }]}>Source URL</Text>
-                                <TextInput
-                   style={[styles.input, {
-                     borderColor: colors.inputBorder,
-                     color: colors.text,
-                   }]}
-                   placeholder="www.cookingwebite.com"
-                   placeholderTextColor={colors.placeholderText}
-                   value={sourceUrl}
-                   onChangeText={setSourceUrl}
-                   autoCapitalize="none"
-                   autoCorrect={false}
-                 />
-
-                <TouchableOpacity
-                  style={[styles.saveButton, { backgroundColor: colors.tint }]}
-                  onPress={handleSave}
-                >
-                  <Text style={[styles.saveButtonText, { color: colors.background }]}>Save Recipe</Text>
-                </TouchableOpacity>
-              </View>
-            </View>
+            <RecipeForm mode="add" onSave={handleSave} />
           </ContentWrapper>
         </NestableScrollContainer>
       </View>
-      <CustomPopup
-        visible={showPopup}
-        title={popupConfig.title}
-        message={popupConfig.message}
-        buttons={popupConfig.buttons}
-        onClose={() => setShowPopup(false)}
-      />
-      <TextInputPopup
-        visible={itemEditor.visible}
-        title={itemEditor.mode === 'add' ? 'Add item' : 'Edit item'}
-        initialValue={itemEditor.initialText}
-        placeholder={itemEditor.section === 'ingredients' ? 'e.g. 200g tomatoes' : 'e.g. Sauté onions until soft'}
-        confirmText={itemEditor.mode === 'add' ? 'Add' : 'Save'}
-        onConfirm={handleConfirmItemEditor}
-        onCancel={() => setItemEditor(prev => ({ ...prev, visible: false }))}
-        initialType={itemEditor.isHeader ? 'header' : 'item'}
-        onDelete={itemEditor.mode === 'edit' && itemEditor.itemId ? (() => {
-          const { section, groupId, itemId } = itemEditor;
-          const groups = findGroupsBySection(section);
-          const idx = groups.findIndex(g => g.id === groupId);
-          if (idx === -1) { setItemEditor(prev => ({ ...prev, visible: false })); return; }
-          const group = groups[idx];
-          const nextItems = group.items.filter(it => it.id !== itemId);
-          const next = [...groups];
-          next[idx] = { ...group, items: nextItems };
-          setGroupsBySection(section, next);
-          setItemEditor(prev => ({ ...prev, visible: false }));
-        }) : undefined}
-        deleteText={'Delete'}
-      />
     </GestureHandlerRootView>
   );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  form: {
-    padding: 16,
-  },
-  input: {
-    borderBottomWidth: 1,
-    borderWidth: 0,
-    borderColor: '#ddd',
-    borderRadius: 0,
-    marginBottom: 16,
-    fontSize: 16,
-    minHeight: 48,
-    padding: 12,
-    textAlignVertical: 'center',
-    lineHeight: 20,
-  },
-  imageButton: {
-    backgroundColor: '#007AFF',
-    padding: 16,
-    borderRadius: 8,
-    alignItems: 'center',
-    marginBottom: 24,
-  },
-  imageButtonText: {
-    color: '#fff',
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  sectionTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    marginBottom: 12,
-    marginTop: 40,
-    opacity: 0.7,
-  },
-  ingredient: {
-    fontSize: 16,
-    flex: 1,
-  },
-    ingredientInput: {
-     flexDirection: 'row',
-     alignItems: 'center',
-     marginBottom: 24,
-   },
-  addButton: {
-    backgroundColor: '#007AFF',
-    width: 48,
-    height: 48,
-    borderRadius: 8,
-    marginLeft: 8,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  instructionsInput: {
-    height: 300,
-    textAlignVertical: 'top',
-  },
-  saveButton: {
-    backgroundColor: '#34C759',
-    padding: 16,
-    borderRadius: 8,
-    alignItems: 'center',
-    marginTop: 24,
-  },
-  saveButtonText: {
-    color: '#fff',
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  imagePreview: {
-    width: '100%',
-    height: 200,
-    marginBottom: 16,
-    borderRadius: 8,
-    overflow: 'hidden',
-    backgroundColor: '#f0f0f0',
-  },
-  previewImage: {
-    width: '100%',
-    height: '100%',
-    resizeMode: 'cover',
-  },
-  ingredientRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    marginBottom: 8,
-  },
-  deleteButton: {
-    backgroundColor: '#8e8e93',
-    width: 46,
-    aspectRatio: 1,
-    borderRadius: 8,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  detailsRow: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginBottom: 16,
-    gap: 8,
-  },
-  detailInput: {
-    flex: 1,
-  },
-  detailLabel: {
-    fontSize: 15,
-    marginBottom: 8,
-    marginTop: 12,
-    opacity: 0.7,
-  },
-  tagsScrollContainer: {
-    overflow: 'visible',
-    marginHorizontal: -16,
-  },
-  tagsContainer: {
-    flexDirection: 'row',
-    gap: 8,
-    overflow: 'visible',
-    paddingHorizontal: 16,
-  },
-  tagContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingVertical: 4,
-    paddingLeft: 12,
-    paddingRight: 4,
-    borderRadius: 16,
-    height: 32,
-    marginBottom: 16,
-  },
-  tagText: {
-    fontSize: 14,
-    marginRight: 4,
-    fontWeight: '500',
-  },
-  tagDeleteButton: {
-    height: 32,
-    width: 32,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  tagDeleteIcon: {
-    opacity: 0.8,
-  },
-  dragHandle: {
-    width: 46,
-    borderRadius: 8,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-}); 
+}
